+++ app/exceptions/boat_pricing_errors.ts
@@ -1,0 +1,7 @@
+export class InvalidPricingRangeError extends Error {
+  name = 'InvalidPricingRangeError'
+
+  constructor() {
+    super('Maximum duration must be greater than or equal to minimum duration')
+  }
+}